pyproject.toml
[tool.poetry]
name = "discord-entry-notification-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
discord = "^1.0.1"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"
pylint = "^2.8.2"
python-dotenv = "^0.17.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
